% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push.R
\name{push.message}
\alias{push.message}
\title{Show a push message}
\usage{
push.message(msg, ...)
}
\arguments{
\item{msg}{the push message}

\item{...}{named options as described here \href{https://pushjs.org/docs/options}{https://pushjs.org/docs/options}}
}
\description{
Show a push message
}
